classical/aml-cli-v2/data-science/src/requirements/feathr-requirements.txt
numpy==1.22.0
pandavro 
# git+https://github.com/feathr-ai/feathr.git#subdirectory=feathr_project&egg=feathr[notebook]
feathr
scikit-learn 
azure-core
ipykernel
pyspark
azure-cli
python-dotenv